Made General Parsing Algorithm
Need to Make Sure an Implicit Unmarshaller is In Scope Which Means utilizing both DefaultJsonProtocol from spray.json and the scaladsl.SprayJsonSupport trait. These two being combined in the jsonParserProtocol should allow us to create representations of our data and then pull those in by declaring which we would like to get.